+++ opentelemetry-user-events-logs/src/logs/mod.rs
@@ -1,4 +1,5 @@
 mod exporter;
+mod hex_buf;
 mod processor;
 
 #[cfg(feature = "experimental_eventname_callback")]